Correct tzfpy docs
It's true that tzfpy is not so accurate around bounders(about 3km ranges).
But it is because of using simplified polygons.

tzfpy uses rectangles as index to speed up but auto fall back to polygon data if cache is missed.
It's written in Rust here: https://github.com/ringsaturn/tzf-rs/blob/v0.4.1/src/lib.rs#L307-L318